requirements.txt
google-generativeai
chromadb
pillow
python-dotenv
pandas
fastapi
uvicorn
